検索・キーワード予約ダイアログに検索条件のプリセットを追加
とりあえず、検索条件のプリセットを選ぶことが出来るだけで、録画プリセットのように
現在のプリセットの表示機能などはありません。

その他一部変更しています。
・検索・キーワード予約ダイアログなどからプリセット設定画面を表示するボタンを追加
・プリセットのデフォルト名の変更を可能とした
・録画プリセットの扱いをやや簡略化

+++ EpgTimer/EpgTimer/Menu/CmdExe.cs
@@ -395,7 +395,7 @@ namespace EpgTimer
 
             if (e.Command == EpgCmds.ChgOnPreset)
             {
-                return MenuUtil.ChangeOnPreset(infoList, (uint)CmdExeUtil.ReadIdData(e, 0, 0xFE));
+                return MenuUtil.ChangeOnPreset(infoList, CmdExeUtil.ReadIdData(e, 0, 0xFE));
             }
             else if (e.Command == EpgCmds.ChgRecmode)
             {
@@ -539,7 +539,7 @@ namespace EpgTimer
                     RecPresetItem pre_0 = listr[0].RecSettingInfo.LookUpPreset(listr[0].IsManual);
                     RecPresetItem value = listr.All(data => data.RecSettingInfo.LookUpPreset(data.IsManual).ID == pre_0.ID) ? pre_0 : null;
                     subMenu.Header = string.Format("プリセット : {0}", value == null ? "*" : value.DisplayName);
-                    SetCheckmarkSubMenus(subMenu, value == null ? int.MinValue : (int)value.ID);
+                    SetCheckmarkSubMenus(subMenu, value == null ? int.MinValue : value.ID);
                 }
                 else if (subMenu.Tag == EpgCmdsEx.ChgResModeMenu)
                 {
